Introduced `plugins.ready` event.
